Relocate Paper module classes for better structure.
