+++ paper/src/main/java/net/thenextlvl/commander/command/HideCommand.java
@@ -1,5 +1,6 @@
-package net.thenextlvl.commander.paper.command;
+package net.thenextlvl.commander.command;
 
+import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
@@ -7,38 +8,37 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.paper.CommanderPlugin;
+import net.thenextlvl.commander.CommanderPlugin;
 import org.bukkit.Bukkit;
-import org.bukkit.command.Command;
 import org.bukkit.entity.Player;
 
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class UnregisterCommand {
+class HideCommand {
     private final CommanderPlugin plugin;
 
-    public ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("unregister")
+    ArgumentBuilder<CommandSourceStack, ?> create() {
+        return Commands.literal("hide")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
                             Bukkit.getCommandMap().getKnownCommands().values().stream()
-                                    .map(Command::getLabel)
-                                    .filter(s -> !plugin.commandRegistry().isUnregistered(s))
+                                    .map(org.bukkit.command.Command::getLabel)
+                                    .filter(s -> !plugin.commandRegistry().isHidden(s))
                                     .filter(s -> s.contains(suggestions.getRemaining()))
                                     .map(StringArgumentType::escapeIfRequired)
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::unregister));
+                        .executes(this::hide));
     }
 
-    private int unregister(CommandContext<CommandSourceStack> context) {
+    private int hide(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var success = plugin.commandRegistry().unregister(command);
-        var message = success ? "command.unregistered" : "nothing.changed";
+        var success = plugin.commandRegistry().hide(command);
+        var message = success ? "command.hidden" : "nothing.changed";
         plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
         if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
-        return com.mojang.brigadier.Command.SINGLE_SUCCESS;
+        return Command.SINGLE_SUCCESS;
     }
 }
