+++ paper/src/main/java/net/thenextlvl/commander/command/PermissionQueryCommand.java
@@ -1,6 +1,5 @@
-package net.thenextlvl.commander.paper.command;
+package net.thenextlvl.commander.command;
 
-import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
@@ -8,38 +7,39 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.paper.CommanderPlugin;
+import net.thenextlvl.commander.CommanderPlugin;
+import org.bukkit.Bukkit;
+import org.bukkit.command.Command;
 
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class ResetCommand {
+class PermissionQueryCommand {
     private final CommanderPlugin plugin;
 
     public ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("reset")
+        return Commands.literal("query")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
-                            plugin.commandRegistry().hiddenCommands().stream()
-                                    .filter(s -> s.contains(suggestions.getRemaining()))
-                                    .map(StringArgumentType::escapeIfRequired)
-                                    .forEach(suggestions::suggest);
-                            plugin.commandRegistry().unregisteredCommands().stream()
+                            Bukkit.getCommandMap().getKnownCommands().values().stream()
+                                    .map(Command::getLabel)
                                     .filter(s -> s.contains(suggestions.getRemaining()))
                                     .map(StringArgumentType::escapeIfRequired)
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::reset));
+                        .executes(this::query));
     }
 
-    private int reset(CommandContext<CommandSourceStack> context) {
+    private int query(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var s1 = plugin.permissionOverride().reset(command);
-        var s3 = plugin.commandRegistry().register(command);
-        var s2 = plugin.commandRegistry().reveal(command);
-        var message = s1 || s2 || s3 ? "command.reset" : "nothing.changed";
-        plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
-        return Command.SINGLE_SUCCESS;
+        var registered = Bukkit.getCommandMap().getKnownCommands().get(command);
+        var permission = registered != null ? registered.getPermission() : null;
+        var message = registered == null ? "command.unknown" : permission != null ?
+                "permission.query.defined" : "permission.query.undefined";
+        plugin.bundle().sendMessage(sender, message,
+                Placeholder.parsed("permission", String.valueOf(permission)),
+                Placeholder.parsed("command", command));
+        return com.mojang.brigadier.Command.SINGLE_SUCCESS;
     }
 }
