+++ paper/src/main/java/net/thenextlvl/commander/command/PermissionResetCommand.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.commander.paper.command;
+package net.thenextlvl.commander.command;
 
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
@@ -8,33 +8,33 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.paper.CommanderPlugin;
+import net.thenextlvl.commander.CommanderPlugin;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class RegisterCommand {
+class PermissionResetCommand {
     private final CommanderPlugin plugin;
 
     public ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("register")
+        return Commands.literal("reset")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
-                            plugin.commandRegistry().unregisteredCommands().stream()
+                            plugin.permissionOverride().originalPermissions().keySet().stream()
                                     .filter(s -> s.contains(suggestions.getRemaining()))
                                     .map(StringArgumentType::escapeIfRequired)
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::register));
+                        .executes(this::reset));
     }
 
-    private int register(CommandContext<CommandSourceStack> context) {
+    private int reset(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var success = plugin.commandRegistry().register(command);
-        var message = success ? "command.registered" : "nothing.changed";
+        var success = plugin.permissionOverride().reset(command);
+        var message = success ? "permission.reset" : "nothing.changed";
         plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
         if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
         return Command.SINGLE_SUCCESS;
