+++ paper/src/main/java/net/thenextlvl/commander/command/RegisterCommand.java
@@ -1,5 +1,6 @@
-package net.thenextlvl.commander.paper.command;
+package net.thenextlvl.commander.command;
 
+import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
@@ -7,39 +8,35 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.paper.CommanderPlugin;
+import net.thenextlvl.commander.CommanderPlugin;
 import org.bukkit.Bukkit;
-import org.bukkit.command.Command;
+import org.bukkit.entity.Player;
 
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class PermissionQueryCommand {
+class RegisterCommand {
     private final CommanderPlugin plugin;
 
     public ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("query")
+        return Commands.literal("register")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
-                            Bukkit.getCommandMap().getKnownCommands().values().stream()
-                                    .map(Command::getLabel)
+                            plugin.commandRegistry().unregisteredCommands().stream()
                                     .filter(s -> s.contains(suggestions.getRemaining()))
                                     .map(StringArgumentType::escapeIfRequired)
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::query));
+                        .executes(this::register));
     }
 
-    private int query(CommandContext<CommandSourceStack> context) {
+    private int register(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var registered = Bukkit.getCommandMap().getKnownCommands().get(command);
-        var permission = registered != null ? registered.getPermission() : null;
-        var message = registered == null ? "command.unknown" : permission != null ?
-                "permission.query.defined" : "permission.query.undefined";
-        plugin.bundle().sendMessage(sender, message,
-                Placeholder.parsed("permission", String.valueOf(permission)),
-                Placeholder.parsed("command", command));
-        return com.mojang.brigadier.Command.SINGLE_SUCCESS;
+        var success = plugin.commandRegistry().register(command);
+        var message = success ? "command.registered" : "nothing.changed";
+        plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
+        if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
+        return Command.SINGLE_SUCCESS;
     }
 }
