+++ paper/src/main/java/net/thenextlvl/commander/command/ResetCommand.java
@@ -1,5 +1,6 @@
-package net.thenextlvl.commander.paper.command;
+package net.thenextlvl.commander.command;
 
+import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
@@ -7,39 +8,38 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.paper.CommanderPlugin;
-import org.bukkit.Bukkit;
-import org.bukkit.command.Command;
-import org.bukkit.entity.Player;
+import net.thenextlvl.commander.CommanderPlugin;
 
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class PermissionUnsetCommand {
+class ResetCommand {
     private final CommanderPlugin plugin;
 
     public ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("unset")
+        return Commands.literal("reset")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
-                            Bukkit.getCommandMap().getKnownCommands().values().stream()
-                                    .map(Command::getLabel)
+                            plugin.commandRegistry().hiddenCommands().stream()
+                                    .filter(s -> s.contains(suggestions.getRemaining()))
+                                    .map(StringArgumentType::escapeIfRequired)
+                                    .forEach(suggestions::suggest);
+                            plugin.commandRegistry().unregisteredCommands().stream()
                                     .filter(s -> s.contains(suggestions.getRemaining()))
                                     .map(StringArgumentType::escapeIfRequired)
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::unset));
+                        .executes(this::reset));
     }
 
-    private int unset(CommandContext<CommandSourceStack> context) {
+    private int reset(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var success = plugin.permissionOverride().override(command, null);
-        var message = success ? "permission.set" : "nothing.changed";
-        plugin.bundle().sendMessage(sender, message,
-                Placeholder.parsed("permission", "null"),
-                Placeholder.parsed("command", command));
-        if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
-        return com.mojang.brigadier.Command.SINGLE_SUCCESS;
+        var s1 = plugin.permissionOverride().reset(command);
+        var s3 = plugin.commandRegistry().register(command);
+        var s2 = plugin.commandRegistry().reveal(command);
+        var message = s1 || s2 || s3 ? "command.reset" : "nothing.changed";
+        plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
+        return Command.SINGLE_SUCCESS;
     }
 }
