+++ paper/src/main/java/net/thenextlvl/commander/command/UnregisterCommand.java
@@ -1,6 +1,5 @@
-package net.thenextlvl.commander.paper.command;
+package net.thenextlvl.commander.command;
 
-import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
@@ -8,37 +7,38 @@ import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.paper.CommanderPlugin;
+import net.thenextlvl.commander.CommanderPlugin;
 import org.bukkit.Bukkit;
+import org.bukkit.command.Command;
 import org.bukkit.entity.Player;
 
 @RequiredArgsConstructor
 @SuppressWarnings("UnstableApiUsage")
-class HideCommand {
+class UnregisterCommand {
     private final CommanderPlugin plugin;
 
-    ArgumentBuilder<CommandSourceStack, ?> create() {
-        return Commands.literal("hide")
+    public ArgumentBuilder<CommandSourceStack, ?> create() {
+        return Commands.literal("unregister")
                 .then(Commands.argument("command", StringArgumentType.string())
                         .suggests((context, suggestions) -> {
                             Bukkit.getCommandMap().getKnownCommands().values().stream()
-                                    .map(org.bukkit.command.Command::getLabel)
-                                    .filter(s -> !plugin.commandRegistry().isHidden(s))
+                                    .map(Command::getLabel)
+                                    .filter(s -> !plugin.commandRegistry().isUnregistered(s))
                                     .filter(s -> s.contains(suggestions.getRemaining()))
                                     .map(StringArgumentType::escapeIfRequired)
                                     .forEach(suggestions::suggest);
                             return suggestions.buildFuture();
                         })
-                        .executes(this::hide));
+                        .executes(this::unregister));
     }
 
-    private int hide(CommandContext<CommandSourceStack> context) {
+    private int unregister(CommandContext<CommandSourceStack> context) {
         var sender = context.getSource().getSender();
         var command = context.getArgument("command", String.class);
-        var success = plugin.commandRegistry().hide(command);
-        var message = success ? "command.hidden" : "nothing.changed";
+        var success = plugin.commandRegistry().unregister(command);
+        var message = success ? "command.unregistered" : "nothing.changed";
         plugin.bundle().sendMessage(sender, message, Placeholder.parsed("command", command));
         if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
-        return Command.SINGLE_SUCCESS;
+        return com.mojang.brigadier.Command.SINGLE_SUCCESS;
     }
 }
